Refactored wrapping table access
