+++ src/data/index.js
@@ -8,7 +8,7 @@
 
 var types = require('../utilities/types'),
     assert = require('../utilities/assert').argument,
-    queries = require('../query');
+    filters = require('./filters');
 
 /**
 Create an instance of the data provider specified in the configuration.
@@ -24,7 +24,7 @@ module.exports = function (configuration) {
 
     var api = function (table, context) {
         assert(table, 'A table was not specified');
-        return wrapTableAccess(provider, table, context);
+        return filters.apply(provider, table, context);
     };
 
     api.execute = function (statement) {
@@ -37,65 +37,6 @@ module.exports = function (configuration) {
     function createProvider() {
         var provider = (configuration && configuration.data && configuration.data.provider) || 'memory';
         return (types.isFunction(provider) ? provider : require('./' + provider))(configuration.data);
-    }
-
-    function wrapTableAccess(provider, table, context) {
-        var tableAccess = provider(table);
-
-        context = context || {};
-
-        return {
-            read: function (query) {
-                return tableAccess.read(applyFilters(query));
-            },
-            update: function (item, query) {
-                assert(item, 'An item to update was not provided');
-                return tableAccess.update(applyTransforms(item), applyFilters(query));
-            },
-            insert: function (item) {
-                assert(item, 'An item to insert was not provided');
-                return tableAccess.insert(applyTransforms(item));
-            },
-            delete: function (query, version) {
-                assert(query, 'The delete query was not provided');
-                return tableAccess.delete(applyFilters(query), version);
-            },
-            undelete: function (query, version) {
-                assert(query, 'The undelete query was not provided');
-                return tableAccess.undelete(applyFilters(query), version);
-            },
-            truncate: function () {
-                return tableAccess.truncate();
-            },
-            initialize: function () {
-                return tableAccess.initialize();
-            },
-            schema: function () {
-                return tableAccess.schema();
-            }
-        };
-
-        function applyFilters(query) {
-            if(!table.filters)
-                return query;
-
-            return table.filters.reduce(function (query, filter) {
-                // this is a bit of trickery to allow filters to either 
-                // modify the existing query or return a different query
-                return filter(query, context) || query;
-            }, query || queries.create(table.name));
-        }
-
-        function applyTransforms(item) {
-            if(!table.transforms)
-                return item;
-
-            return table.transforms.reduce(function (item, transform) {
-                // same trick as in filters
-                // see the test in data/integration/filters for an example                
-                return transform(item, context) || item;
-            }, item);
-        }
     }
 };
